+++ R/mixedBayes.R
@@ -7,12 +7,12 @@
 #' @param X the long format matrix of the intercept and time effects (time effects are optional).
 #' @param w the long format matrix of interactions between genetic factors and environment (treatment) factors.
 #' @param k the number of repeated measurements.
-#' @param iterations the number of MCMC iterations.
-#' @param burn.in the number of iterations for burn-in.
-#' @param slope logical flag. If TRUE, random intercept-and-slope model will be used.
-#' @param robust logical flag. If TRUE, robust methods will be used.
-#' @param quant specify different quantiles when applying robust methods.
-#' @param sparse logical flag. If TRUE, spike-and-slab priors will be used to shrink coefficients of irrelevant covariates to zero exactly.
+#' @param iterations the number of MCMC iterations. The default value is 10,000.
+#' @param burn.in the number of iterations for burn-in. If NULL, the first half of MCMC iterations will be used as burn-ins.
+#' @param slope logical flag. If TRUE, random intercept-and-slope model will be used. Otherwise, random intercept model will be used. The default value is TRUE.
+#' @param robust logical flag. If TRUE, robust methods will be used. Otherwise, non-robust methods will be used. The default value is TRUE.
+#' @param quant the quantile level specified by users. The default value is 0.5.
+#' @param sparse logical flag. If TRUE, spike-and-slab priors will be adopted to impose exact sparsity on regression coefficients. Otherwise, Laplacian shrinkage will be adopted. The default value is TRUE.
 #' @param structure two choices are available. "bi-level" for selection on both the main and interaction effects corresponding to individual and group levels. "individual" for selections on individual-level only.
 #' @return an object of class `mixedBayes' is returned, which is a list with component:
 #' \item{posterior}{the posteriors of coefficients.}
@@ -21,8 +21,11 @@
 #' \item{iterations}{the total number of iterations.}
 #'
 #' @details Consider the data model described in "\code{\link{data}}":
-#' \deqn{Y_{ij} = X_{ij}^\top\gamma_{0}+E_{ij}^\top\gamma_{1}+\sum_{l=1}^{p}G_{ijl}\gamma_{2l}+\sum_{l=1}^{p}W_{ijl}^\top\gamma_{3l}+Z_{ij}^\top\alpha_{i}+\epsilon_{ij}.}
-#' where \eqn{\gamma_{2l}} is the main effect of the \eqn{l}th genetic variant. The interaction effects is corresponding to the coefficient vector \eqn{\gamma_{3l}=(\gamma_{3l1}, \gamma_{3l2},\ldots,\gamma_{3lm})^\top}, with \eqn{W_{ij} = G_{ij}\bigotimes E_{ij}}.
+#' \deqn{Y_{ij} = X_{ij}^\top\gamma_{0}+E_{ij}^\top\gamma_{1}+\sum_{l=1}^{p}G_{ijl}\gamma_{2l}+\sum_{l=1}^{p}W_{ijl}^\top\gamma_{3l}+Z_{ij}^\top\alpha_{i}+\epsilon_{ij}.}, with \eqn{W_{ij} = G_{ij}\bigotimes E_{ij}}.
+#'
+#' where \eqn{\gamma_{0}} is the coefficient vector for \eqn{X_{ij}}, \eqn{\gamma_{1}} is the coefficient vector for \eqn{E_{ij}}, \eqn{\gamma_{2l}} is the coefficient for the main effect of the \eqn{l}th genetic variant, and \eqn{\gamma_{3l}} is the coefficient vector for the interaction effect of the \eqn{l}th genetic variant with environment factors.
+#'
+#' For random intercept-and-slope model, \eqn{Z_{ij}^\top = (1,j)} and  \eqn{\alpha_{i} = (\alpha_{i1},\alpha_{i2})^\top}. For random intercept model, \eqn{Z_{ij}^\top = 1} and \eqn{\alpha_{i} = \alpha_{i1}}.
 #'
 #' When `structure="bi-level"`, bi-level selection will be conducted. If `structure="individual"`, individual-level selection will be conducted.
 #'
@@ -104,11 +107,14 @@ mixedBayes <- function(y,e,X,g,w,k, iterations=10000, burn.in=NULL, slope=TRUE, 
     }
   }
 
+  coeff.treatment = apply(out$GS.alpha[-(1:BI),,drop=FALSE], 2, stats::median);
   coeff.main = apply(out$GS.beta[-(1:BI),,drop=FALSE], 2, stats::median);
   coeff.interaction = apply(out$GS.eta[-(1:BI),,drop=FALSE], 2, stats::median);
+  coeff.random = apply(out$GS.ata[-(1:BI),,drop=FALSE], 2, stats::median);
+  coeff.random = matrix(coeff.random,ncol = length(y)/k)
 
 
-  coefficient = list(main=coeff.main, inter=coeff.interaction)
+  coefficient = list(treatment = coeff.treatment, main=coeff.main, interaction=coeff.interaction, random=coeff.random)
 
   fit = list(posterior = out, coefficient=coefficient, burn.in = BI, iterations=iterations)